requirements.txt
pandas
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
filecache
openpyxl
xlrd
telethon==1.16.0
phonenumbers==8.12.9
python-dotenv==0.10.3
